시간 랭킹에서 시간을 Double로 변경

+++ Leaderboard/Domain/Ranking.swift
@@ -21,5 +21,5 @@ struct ScoreRanking: Ranking {
 struct TimeRanking: Ranking {
     let userId: String
     let nickname: String
-    let seconds: Int
+    let seconds: Double
 }
